refactor AnimalProfile to extract FosterHistory table - Create FosterHistoryTable component
